Split mock pull request planner
